Added movie controller and updated seed file and db

+++ db/migrate/20170724150530_create_parks.rb
@@ -2,7 +2,7 @@ class CreateParks < ActiveRecord::Migration[5.1]
   def change
     create_table :parks do |t|
       t.string :name
-      t.integer :location
+      t.float :location
       t.string :neighborhood
       t.string :park_url
